Crate androidpublisher2

Features

Handle the following Resources with ease from the central hub ...

edits (apklistings.delete, apklistings.deleteall, apklistings.get, apklistings.list, apklistings.patch, apklistings.update, apks.addexternallyhosted, apks.list, apks.upload, commit, delete, details.get, details.patch, details.update, expansionfiles.get, expansionfiles.patch, expansionfiles.update, expansionfiles.upload, get, images.delete, images.deleteall, images.list, images.upload, insert, listings.delete, listings.deleteall, listings.get, listings.list, listings.patch, listings.update, testers.get, testers.patch, testers.update, tracks.get, tracks.list, tracks.patch, tracks.update and validate)

inappproducts (batch, delete, get, insert, list, patch and update)

purchases (products.get, subscriptions.cancel, subscriptions.defer, subscriptions.get, subscriptions.refund and subscriptions.revoke)

Everything else about the Android Publisher API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.inappproducts().patch(...).doit()
let r = hub.inappproducts().insert(...).doit()
let r = hub.inappproducts().get(...).doit()
let r = hub.inappproducts().update(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
androidpublisher2 = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate androidpublisher2;
use androidpublisher2::InAppProduct;
use androidpublisher2::cmn::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use androidpublisher2::AndroidPublisher;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AndroidPublisher::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req: InAppProduct = Default::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.inappproducts().patch(&req, "packageName", "sku")
             .auto_convert_missing_prices(true)
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(value) => println!("Result Value: {:?}", value),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Macros

map!
This macro is advertised in the documentation, which is why we deliver it as well

Structs

AndroidPublisher
Central instance to access all AndroidPublisher related resource activities

Apk
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

apks.upload edits (response)

ApkBinary
Represents the binary payload of an APK.

This type is not used in any activity, and only used as part of another schema.

ApkListing
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

apklistings.get edits (response)

apklistings.patch edits (request|response)

apklistings.update edits (request|response)

ApkListingsListResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

apklistings.list edits (response)

ApksAddExternallyHostedRequest
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

apks.addexternallyhosted edits (request)

ApksAddExternallyHostedResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

apks.addexternallyhosted edits (response)

ApksListResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

apks.list edits (response)

AppDetails
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

details.patch edits (request|response)

details.update edits (request|response)

details.get edits (response)

AppEdit
Represents an edit of an app. An edit allows clients to make multiple changes before committing them in one operation.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert edits (request|response)

commit edits (response)

validate edits (response)

get edits (response)

EditApkAddexternallyhostedMethodBuilder
Creates a new APK without uploading the APK itself to Google Play, instead hosting the APK at a specified URL. This function is only available to enterprises using Google Play for work whose application is configured to restrict distribution to the enterprise domain.

EditApkListMethodBuilder
A builder for the apks.list method supported by a edit resource. It is not used directly, but through a EditMethodsBuilder.

EditApkUploadMethodBuilder
A builder for the apks.upload method supported by a edit resource. It is not used directly, but through a EditMethodsBuilder.

EditApklistingDeleteMethodBuilder
Deletes the APK-specific localized listing for a specified APK and language code.

EditApklistingDeleteallMethodBuilder
Deletes all the APK-specific localized listings for a specified APK.

EditApklistingGetMethodBuilder
Fetches the APK-specific localized listing for a specified APK and language code.

EditApklistingListMethodBuilder
Lists all the APK-specific localized listings for a specified APK.

EditApklistingPatchMethodBuilder
Updates or creates the APK-specific localized listing for a specified APK and language code. This method supports patch semantics.

EditApklistingUpdateMethodBuilder
Updates or creates the APK-specific localized listing for a specified APK and language code.

EditCommitMethodBuilder
Commits/applies the changes made in this edit back to the app.

EditDeleteMethodBuilder
Deletes an edit for an app. Creating a new edit will automatically delete any of your previous edits so this method need only be called if you want to preemptively abandon an edit.

EditDetailGetMethodBuilder
Fetches app details for this edit. This includes the default language and developer support contact information.

EditDetailPatchMethodBuilder
Updates app details for this edit. This method supports patch semantics.

EditDetailUpdateMethodBuilder
Updates app details for this edit.

EditExpansionfileGetMethodBuilder
Fetches the Expansion File configuration for the APK specified.

EditExpansionfilePatchMethodBuilder
Updates the APK's Expansion File configuration to reference another APK's Expansion Files. To add a new Expansion File use the Upload method. This method supports patch semantics.

EditExpansionfileUpdateMethodBuilder
Updates the APK's Expansion File configuration to reference another APK's Expansion Files. To add a new Expansion File use the Upload method.

EditExpansionfileUploadMethodBuilder
Uploads and attaches a new Expansion File to the APK specified.

EditGetMethodBuilder
Returns information about the edit specified. Calls will fail if the edit is no long active (e.g. has been deleted, superseded or expired).

EditImageDeleteMethodBuilder
Deletes the image (specified by id) from the edit.

EditImageDeleteallMethodBuilder
Deletes all images for the specified language and image type.

EditImageListMethodBuilder
Lists all images for the specified language and image type.

EditImageUploadMethodBuilder
Uploads a new image and adds it to the list of images for the specified language and image type.

EditInsertMethodBuilder
Creates a new edit for an app, populated with the app's current state.

EditListingDeleteMethodBuilder
Deletes the specified localized store listing from an edit.

EditListingDeleteallMethodBuilder
Deletes all localized listings from an edit.

EditListingGetMethodBuilder
Fetches information about a localized store listing.

EditListingListMethodBuilder
Returns all of the localized store listings attached to this edit.

EditListingPatchMethodBuilder
Creates or updates a localized store listing. This method supports patch semantics.

EditListingUpdateMethodBuilder
Creates or updates a localized store listing.

EditMethodsBuilder
A builder providing access to all methods supported on edit resources. It is not used directly, but through the AndroidPublisher hub.

EditTesterGetMethodBuilder
A builder for the testers.get method supported by a edit resource. It is not used directly, but through a EditMethodsBuilder.

EditTesterPatchMethodBuilder
A builder for the testers.patch method supported by a edit resource. It is not used directly, but through a EditMethodsBuilder.

EditTesterUpdateMethodBuilder
A builder for the testers.update method supported by a edit resource. It is not used directly, but through a EditMethodsBuilder.

EditTrackGetMethodBuilder
Fetches the track configuration for the specified track type. Includes the APK version codes that are in this track.

EditTrackListMethodBuilder
Lists all the track configurations for this edit.

EditTrackPatchMethodBuilder
Updates the track configuration for the specified track type. When halted, the rollout track cannot be updated without adding new APKs, and adding new APKs will cause it to resume. This method supports patch semantics.

EditTrackUpdateMethodBuilder
Updates the track configuration for the specified track type. When halted, the rollout track cannot be updated without adding new APKs, and adding new APKs will cause it to resume.

EditValidateMethodBuilder
Checks that the edit can be successfully committed. The edit's changes are not applied to the live app.

ExpansionFile
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

expansionfiles.get edits (response)

expansionfiles.patch edits (request|response)

expansionfiles.update edits (request|response)

ExpansionFilesUploadResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

expansionfiles.upload edits (response)

ExternallyHostedApk
Defines an APK available for this application that is hosted externally and not uploaded to Google Play. This function is only available to enterprises who are using Android for Work, and whos application is restricted to the enterprise private channel

This type is not used in any activity, and only used as part of another schema.

ExternallyHostedApkUsesPermission
A permission used by this APK.

This type is not used in any activity, and only used as part of another schema.

Image
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

ImagesDeleteAllResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

images.deleteall edits (response)

ImagesListResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

images.list edits (response)

ImagesUploadResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

images.upload edits (response)

InAppProduct
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

patch inappproducts (request|response)

insert inappproducts (request|response)

get inappproducts (response)

update inappproducts (request|response)

InAppProductListing
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

InappproductBatchMethodBuilder
A builder for the batch method supported by a inappproduct resource. It is not used directly, but through a InappproductMethodsBuilder.

InappproductDeleteMethodBuilder
Delete an in-app product for an app.

InappproductGetMethodBuilder
Returns information about the in-app product specified.

InappproductInsertMethodBuilder
Creates a new in-app product for an app.

InappproductListMethodBuilder
List all the in-app products for an Android app, both subscriptions and managed in-app products..

InappproductMethodsBuilder
A builder providing access to all methods supported on inappproduct resources. It is not used directly, but through the AndroidPublisher hub.

InappproductPatchMethodBuilder
Updates the details of an in-app product. This method supports patch semantics.

InappproductUpdateMethodBuilder
Updates the details of an in-app product.

InappproductsBatchRequest
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

batch inappproducts (request)

InappproductsBatchRequestEntry
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

InappproductsBatchResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

batch inappproducts (response)

InappproductsBatchResponseEntry
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

InappproductsInsertRequest
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

InappproductsInsertResponse
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

InappproductsListResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list inappproducts (response)

InappproductsUpdateRequest
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

InappproductsUpdateResponse
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Listing
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

listings.update edits (request|response)

listings.patch edits (request|response)

listings.get edits (response)

ListingsListResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

listings.list edits (response)

MonthDay
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

PageInfo
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Price
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

ProductPurchase
A ProductPurchase resource indicates the status of a user's inapp product purchase.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

products.get purchases (response)

PurchaseMethodsBuilder
A builder providing access to all methods supported on purchase resources. It is not used directly, but through the AndroidPublisher hub.

PurchaseProductGetMethodBuilder
Checks the purchase and consumption status of an inapp item.

PurchaseSubscriptionCancelMethodBuilder
Cancels a user's subscription purchase. The subscription remains valid until its expiration time.

PurchaseSubscriptionDeferMethodBuilder
Defers a user's subscription purchase until a specified future expiration time.

PurchaseSubscriptionGetMethodBuilder
Checks whether a user's subscription purchase is valid and returns its expiry time.

PurchaseSubscriptionRefundMethodBuilder
Refunds a user's subscription purchase, but the subscription remains valid until its expiration time and it will continue to recur.

PurchaseSubscriptionRevokeMethodBuilder
Refunds and immediately revokes a user's subscription purchase. Access to the subscription will be terminated immediately and it will stop recurring.

Season
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

SubscriptionDeferralInfo
A SubscriptionDeferralInfo contains the data needed to defer a subscription purchase to a future expiry time.

This type is not used in any activity, and only used as part of another schema.

SubscriptionPurchase
A SubscriptionPurchase resource indicates the status of a user's subscription purchase.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

subscriptions.get purchases (response)

SubscriptionPurchasesDeferRequest
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

subscriptions.defer purchases (request)

SubscriptionPurchasesDeferResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

subscriptions.defer purchases (response)

Testers
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

testers.get edits (response)

testers.patch edits (request|response)

testers.update edits (request|response)

TokenPagination
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Track
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

tracks.update edits (request|response)

tracks.patch edits (request|response)

tracks.get edits (response)

TracksListResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

tracks.list edits (response)

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.