Struct androidpublisher2::AppEdit

pub struct AppEdit {
    pub expiry_time_seconds: Option<String>,
    pub id: Option<String>,
}

Represents an edit of an app. An edit allows clients to make multiple changes before committing them in one operation.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert edits (request|response)

commit edits (response)

validate edits (response)

get edits (response)

Fields

expiry_time_seconds

The time at which the edit will expire and will be no longer valid for use in any subsequent API calls (encoded as seconds since the Epoch).

id

The ID of the edit that can be used in subsequent API calls.